# [optimization] improving pipeline entrypoint performance