### What it does

* improves the ``ClientDecoder`` interface to make decoders implementations reusable
* implements the ``jsonDecoder`` and ``msgpackDecoder`` structs to keep references to the internal buffers and decoders; in this way decoders are reusable multiple times
* ``readAll`` re-implementation that overwrites the given buffer without any kind of allocations (the built-in version allocates a new resizable ``Buffer`` and returns a slice descriptor for each request)
* when a decoder is created with the internal ``Buffer``, the buffer has ``minBufferSize`` size. This improves a lot performances because during each request the buffer is resized many times to contain the ``req.Body``
* the request handler hot path, doesn't make useless allocations when a request is handled without errors.

#### Performance impact
The benchmark includes the trace handling for a single trace with a single span, so the performance gain is linear with the number of traces. Furthermore, because we're not allocating anything (obviously with the exception of the ``[]Trace`` and the built-in server objects), we're reducing the number of GC.

```
benchmark                   old ns/op     new ns/op     delta
BenchmarkHandleTraces-4     13198         9723          -26.33%

benchmark                   old allocs     new allocs     delta
BenchmarkHandleTraces-4     33             23             -30.30%

benchmark                   old bytes     new bytes     delta
BenchmarkHandleTraces-4     6530          1408          -78.44%
```